+++ nextgen/project-expander/src/project_expander.rs
@@ -1,4 +1,4 @@
-use crate::expander_context::{ExpanderContext, ExpansionBoundaries};
+use crate::expander_context::ExpanderContext;
 use crate::tasks_expander::TasksExpander;
 use moon_common::color;
 use moon_config::DependencyConfig;
@@ -17,7 +17,7 @@ impl<'graph, 'query> ProjectExpander<'graph, 'query> {
         Self { context }
     }
 
-    pub fn expand(&mut self, boundaries: &mut ExpansionBoundaries) -> miette::Result<Project> {
+    pub fn expand(&mut self) -> miette::Result<Project> {
         // Clone before expanding!
         let mut project = self.context.project.to_owned();
 
@@ -28,7 +28,7 @@ impl<'graph, 'query> ProjectExpander<'graph, 'query> {
         );
 
         self.expand_deps(&mut project)?;
-        self.expand_tasks(&mut project, boundaries)?;
+        self.expand_tasks(&mut project)?;
 
         Ok(project)
     }
@@ -59,11 +59,7 @@ impl<'graph, 'query> ProjectExpander<'graph, 'query> {
         Ok(())
     }
 
-    pub fn expand_tasks(
-        &mut self,
-        project: &mut Project,
-        boundaries: &mut ExpansionBoundaries,
-    ) -> miette::Result<()> {
+    pub fn expand_tasks(&mut self, project: &mut Project) -> miette::Result<()> {
         let mut tasks = BTreeMap::new();
         let mut expander = TasksExpander::new(&self.context);
 
@@ -78,7 +74,7 @@ impl<'graph, 'query> ProjectExpander<'graph, 'query> {
             expander.expand_env(&mut task)?;
             expander.expand_deps(&mut task)?;
             expander.expand_inputs(&mut task)?;
-            expander.expand_outputs(&mut task, boundaries)?;
+            expander.expand_outputs(&mut task)?;
             expander.expand_args(&mut task)?;
             expander.expand_command(&mut task)?;